feat: add API versioning and public/private URL routing; refactor imports for consistency
