+++ dashboard/api/serializers/provider_serializers.py
@@ -1,15 +1,15 @@
 from rest_framework import serializers
 from datetime import date
-from ..models import Patient
-from ..utils import calculate_age
+from dashboard.models import Provider
+from dashboard.utils import calculate_age
 
 
-class PublicPatientSerializer(serializers.ModelSerializer):
+class PublicProviderSerializer(serializers.ModelSerializer):
     year_of_birth = serializers.SerializerMethodField()
 
     class Meta:
-        model = Patient
-        fields = ['id', 'patient_id', 'year_of_birth',
+        model = Provider
+        fields = ['id', 'provider_id', 'year_of_birth',
                   'sex', 'race', 'ethnicity']
 
     def get_year_of_birth(self, instance):
@@ -28,12 +28,12 @@ class PublicPatientSerializer(serializers.ModelSerializer):
         return rep
 
 
-class PatientSerializer(serializers.ModelSerializer):
+class ProviderSerializer(serializers.ModelSerializer):
     year_of_birth = serializers.SerializerMethodField()
 
     class Meta:
-        model = Patient
-        fields = ['id', 'patient_id', 'year_of_birth',
+        model = Provider
+        fields = ['id', 'provider_id', 'year_of_birth',
                   'sex', 'race', 'ethnicity']
 
     def get_year_of_birth(self, instance):
