+++ dashboard/api/serializers/source_and_department_serializers.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
-from ..models import EncounterSource, Department
+from dashboard.models import EncounterSource, Department
 
 
 class EncounterSourceSerializer(serializers.ModelSerializer):
