+++ dashboard/api/urls/v1/auth_urls.py
@@ -1,0 +1,10 @@
+from django.urls import path
+from dashboard.api.views.auth_views import CustomTokenObtainPairView, LogoutView
+from rest_framework_simplejwt.views import TokenRefreshView, TokenVerifyView
+
+urlpatterns = [
+    path('token/', CustomTokenObtainPairView.as_view(), name='token_obtain_pair'),
+    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
+    path('token/verify/', TokenVerifyView.as_view(), name='token_verify'),
+    path('logout/', LogoutView.as_view(), name='logout'),
+]
