+++ dashboard/api/urls/v1/profile_urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from dashboard.api.views.profile_views import ProfileView
+
+urlpatterns = [
+    path('', ProfileView.as_view(), name='api_profile'),
+]
