+++ dashboard/api/viewsets/public/__init__.py
@@ -1,0 +1,1 @@
+from .public_views import PublicPatientViewSet, PublicProviderViewSet, PublicEncounterViewSet, PublicDepartmentViewSet, PublicMultiModalDataViewSet, PublicEncounterSourceViewSet
