+++ dashboard/signals/encounter_signals.py
@@ -1,6 +1,6 @@
 from django.db.models.signals import post_save
 from django.dispatch import receiver
-from ..models import Encounter, MultiModalData, EncounterFile
+from dashboard.models import Encounter, MultiModalData, EncounterFile
 
 
 @receiver(post_save, sender=Encounter)
