+++ dashboard/signals/profile_signals.py
@@ -1,7 +1,7 @@
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 from django.contrib.auth.models import User
-from ..models import Profile
+from dashboard.models import Profile
 
 
 @receiver(post_save, sender=User)
